fix sul movimento palyer e camera, aggiunto background di prova

+++ src/main/java/it/unibo/bioassault/model/GameObject.java
@@ -6,7 +6,7 @@ import java.awt.*;
 
 public abstract class GameObject {
 
-    protected int x, y; //posizione dell'oggetto
+    protected float x, y; //posizione dell'oggetto
     protected float velX = 0, velY = 0; //velocità oggetto
     protected ID id;
 
@@ -29,23 +29,21 @@ public abstract class GameObject {
     public void setId(ID id) {
         this.id = id;
     }
-
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
-
     public float getVelX() {
         return velX;
     }
